generator/requirements.txt
azure-eventhub==5.3.1
Jinja2==2.11.3